backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
duckdb>=0.9.0
pandas>=2.1.0
pyarrow>=14.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
